+++ src/app/api/models/task.ts
@@ -81,6 +81,8 @@ export class Task extends Entity {
 
   targetStartDate: Date;
   targetDueDate: Date;
+  movedToDiscussAt: Date;
+  discussTimeoutExpiryAt: Date;
 
   public topWeight: number = 0;
   public readonly commentCache: EntityCache<TaskComment> = new EntityCache<TaskComment>();