+++ src/app/api/services/authentication.service.ts
@@ -313,4 +313,12 @@ export class AuthenticationService {
       }),
     );
   }
+
+  public getContentToken(): Observable<string> {
+    return this.httpClient.get(this.AUTH_URL + '/content').pipe(
+      map((response) => {
+        return response['content_auth_token'];
+      }),
+    );
+  }
 }